refactor: split gorse cli admin and context code
